pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-team-platform"
version = "0.1.0"
description = "AI-Powered Workforce Platform - AI agents that replace company roles"
requires-python = ">=3.11"
dependencies = [
    # AI / LLM — supports both OpenAI and Anthropic
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "openai>=1.50.0",
    "anthropic>=0.39.0",
    "tiktoken>=0.7.0",
    # API framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1.0",
    # Data validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Cache
    "redis>=5.1.0",
    # Vector DB
    "chromadb>=0.5.0",
    # Document processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    # QA pipeline
    "textstat>=0.7.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
    "pyjwt[crypto]>=2.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "black>=24.0.0",
    "ipython>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools]
packages = ["agents", "analytics", "api", "core", "db", "frontend", "integrations", "prompts", "ui"]